* Update for ovl v4 and v5 support (btbl still broken) * License header update for bin2c
git-svn-id: https://svn.code.sf.net/p/rct3/code/trunk@139 68d9b628-aa21-0410-a705-e462ad5a1edf

+++ RCT3 Importer/src/libOVLDump/ovldumperstructs.cpp
@@ -35,7 +35,7 @@ void OvlSymbol::SetData(unsigned long data) {
     if (IsPointer)
         throw EOvlD("Tried to set data for a pointer symbol");
 
-    orgsymbol->data = reinterpret_cast<unsigned long *>(data);
+    orgsymbol->data = reinterpret_cast<uint32_t *>(data);
 }
 
 void OvlStringTableEntry::SetString(const char* newstr) const {
